+++ CHANGELOG.md
@@ -4,7 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). See [MAINTAINERS.md](./MAINTAINERS.md)
 for instructions to keep up to date.
 
-## Unreleased
+## Unreleased (target: v0.10.0 release)
+
+### Changed
+
+* Renamed common `atm` 4 flags to `blocks-cache`: 
+  `--common-blocks-cache-{enabled|dir|max-recent-entry-bytes|max-entry-by-age-bytes}`
 
 #### Fixed